requirements.txt
# 基础深度学习库
torch==1.12.1+cu113  # PyTorch，指定版本与CUDA兼容（根据需要修改为合适的版本）
torchvision==0.13.1+cu113  # PyTorch的计算机视觉库
torchaudio==0.12.1  # PyTorch的音频处理库（如果有使用音频数据）

# 科学计算库
numpy==1.23.2  # 数值计算基础库
scipy==1.9.1  # 科学计算，常用于信号处理

# 数据处理和分析
pandas==1.4.3  # 数据处理库
scikit-learn==1.1.2  # 机器学习库，用于模型评估和预处理

# 可视化
matplotlib==3.5.2  # 数据可视化库
seaborn==0.11.2  # 高级数据可视化库，基于matplotlib

# 其他常用库
pyyaml==6.0  # 用于加载yaml配置文件
tqdm==4.64.0  # 用于显示训练进度条
scikit-image==0.19.3  # 图像处理库（可选，根据需要）

# 强化学习和优化
gym==0.26.0  # 强化学习环境库（如果涉及强化学习优化）
tensorboard==2.10.0  # TensorBoard，PyTorch训练可视化工具（可选）

# 小波变换
pywt==1.3.0  # 小波变换库，用于时频域特征提取

# 用于实验的其他库（如文件操作）
shutil==1.0.0  # 文件操作库，用于目录和文件管理

# 版本管理工具
setuptools==62.3.4  # 用于安装和打包Python项目的工具
